Revert QRCodeGeneratorApp.kt
